core Config (bugfix): Allows forcing domain and protocol from CLI

+++ core/Config/Model/Event/YamlSettingEventListener.class.php
@@ -82,11 +82,14 @@ class YamlSettingEventListener extends \Cx\Core\Event\Model\Entity\DefaultEventL
                             if ($_CONFIG['forceProtocolFrontend'] != 'none') {
                                 $protocol = $_CONFIG['forceProtocolFrontend'];
                             }
-                            if (\Cx\Core\Config\Controller\Config::checkAccessibility($protocol, $domainUrl)) {
-                                $this->getComponent('Cache')->deleteNonPagePageCache();
-                            } else {
+                            if (
+                                php_sapi_name() != 'cli' &&
+                                !\Cx\Core\Config\Controller\Config::checkAccessibility($protocol, $domainUrl)
+                            ) {
                                 \Message::add(sprintf($_ARRAYLANG['TXT_CONFIG_UNABLE_TO_SET_MAINDOMAIN'], $domainUrl), \Message::CLASS_ERROR);
                                 $objSetting->setValue($_CONFIG['mainDomainId']);
+                            } else {
+                                $this->getComponent('Cache')->deleteNonPagePageCache();
                             }
                         }
                     }
@@ -94,7 +97,10 @@ class YamlSettingEventListener extends \Cx\Core\Event\Model\Entity\DefaultEventL
 
                 case 'forceProtocolFrontend':
                     if ($_CONFIG['forceProtocolFrontend'] != $value) {
-                        if (!\Cx\Core\Config\Controller\Config::checkAccessibility($value)) {
+                        if (
+                            php_sapi_name() != 'cli' &&
+                            !\Cx\Core\Config\Controller\Config::checkAccessibility($value)
+                        ) {
                             \Message::add($_ARRAYLANG['TXT_CONFIG_UNABLE_TO_SET_PROTOCOL'], \Message::CLASS_ERROR);
                             $objSetting->setValue('none');
                         }
@@ -106,7 +112,10 @@ class YamlSettingEventListener extends \Cx\Core\Event\Model\Entity\DefaultEventL
 
                 case 'forceProtocolBackend':
                     if ($_CONFIG['forceProtocolBackend'] != $value) {
-                        if (!\Cx\Core\Config\Controller\Config::checkAccessibility($value)) {
+                        if (
+                            php_sapi_name() != 'cli' &&
+                            !\Cx\Core\Config\Controller\Config::checkAccessibility($value)
+                        ) {
                             \Message::add($_ARRAYLANG['TXT_CONFIG_UNABLE_TO_SET_PROTOCOL'], \Message::CLASS_ERROR);
                             $objSetting->setValue('none');
                         }
@@ -119,7 +128,10 @@ class YamlSettingEventListener extends \Cx\Core\Event\Model\Entity\DefaultEventL
                     if ($useHttps == 'https') {
                         $protocol = 'https';
                     }
-                    if (!\Cx\Core\Config\Controller\Config::checkAccessibility($protocol)) {
+                    if (
+                        php_sapi_name() != 'cli' &&
+                        !\Cx\Core\Config\Controller\Config::checkAccessibility($protocol)
+                    ) {
                         \Message::add($_ARRAYLANG['TXT_CONFIG_UNABLE_TO_FORCE_MAINDOMAIN'], \Message::CLASS_ERROR);
                         $objSetting->setValue('off');
                     }
